+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Create karpenter custom resources in workload clusters.
+
 ## [0.20.0] - 2025-06-23
 
 ### Changed